refactor: remove Interfaces from AppNavigator and AppController
